drobe/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
coverage==7.4.4
crispy-bootstrap4==2023.1
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-heroku==0.3.1
django-paypal==2.0
djangorestframework==3.15.1
executing==2.0.1
factory-boy==3.3.0
Faker==24.4.0
git-filter-repo==2.38.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
mysqlclient==2.2.1
oauth2==1.9.0.post1
outcome==1.3.0.post0
packaging==24.0
parso==0.8.3
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
requests==2.27.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
whitenoise==6.6.0
wsproto==1.2.0
